Create a generic one roster class that infinitecampus inherits from
